+++ PowerSync/PowerSync.Common/Client/Sync/Stream/StreamingSyncImplementation.cs
@@ -50,8 +50,13 @@ public class StreamingSyncImplementationOptions : AdditionalConnectionOptions
     public ILogger? Logger { get; init; }
 }
 
-public class BaseConnectionOptions(Dictionary<string, object>? parameters = null)
+public class BaseConnectionOptions(Dictionary<string, object>? parameters = null, Dictionary<string, string>? appMetadata = null)
 {
+    /// <summary>
+    /// A set of metadata to be included in service logs.
+    /// </summary>
+    public Dictionary<string, string>? AppMetadata { get; set; } = appMetadata;
+
     /// <summary>
     /// These parameters are passed to the sync rules and will be available under the `user_parameters` object.
     /// </summary>
@@ -60,6 +65,9 @@ public class BaseConnectionOptions(Dictionary<string, object>? parameters = null
 
 public class RequiredPowerSyncConnectionOptions : BaseConnectionOptions
 {
+
+    public new Dictionary<string, string> AppMetadata { get; set; } = new();
+
     public new Dictionary<string, object> Params { get; set; } = new();
 }
 
@@ -97,6 +105,7 @@ public class StreamingSyncImplementation : EventStream<StreamingSyncImplementati
 {
     public static RequiredPowerSyncConnectionOptions DEFAULT_STREAM_CONNECTION_OPTIONS = new()
     {
+        AppMetadata = [],
         Params = [],
     };
 
@@ -381,6 +390,7 @@ public class StreamingSyncImplementation : EventStream<StreamingSyncImplementati
             {
                 var resolvedOptions = new RequiredPowerSyncConnectionOptions
                 {
+                    AppMetadata = options?.AppMetadata ?? DEFAULT_STREAM_CONNECTION_OPTIONS.AppMetadata,
                     Params = options?.Params ?? DEFAULT_STREAM_CONNECTION_OPTIONS.Params,
                 };
 
@@ -550,7 +560,7 @@ public class StreamingSyncImplementation : EventStream<StreamingSyncImplementati
 
         try
         {
-            await Control(PowerSyncControlCommand.START, JsonConvert.SerializeObject(new { parameters = resolvedOptions.Params }));
+            await Control(PowerSyncControlCommand.START, JsonConvert.SerializeObject(new { parameters = resolvedOptions.Params, app_metadata = resolvedOptions.AppMetadata }));
 
             notifyCompletedUploads = () =>
             {