+++ PowerSync/PowerSync.Common/CHANGELOG.md
@@ -9,6 +9,20 @@
 - Reporting progress information about downloaded rows. Sync progress is available through `SyncStatus.DownloadProgress()`.
 - Support bucket priorities.
 - Report `PriorityStatusEntries` on `SyncStatus`.
+- Added ability to specify `AppMetadata` for sync/stream requests.
+
+Note: This requires a PowerSync service version `>=1.17.0` in order for logs to display metadata.
+
+```csharp
+db.Connect(connector, new PowerSync.Common.Client.Sync.Stream.PowerSyncConnectionOptions
+{
+    // This will be included in PowerSync service logs
+    AppMetadata = new Dictionary<string, string>
+    {
+        { "app_version", myAppVersion },
+    }
+});
+```
 
 ## 0.0.5-alpha.1
 - Using the latest version (0.4.9) of the core extension, it introduces support for the Rust Sync implementation and also makes it the default - users can still opt out and use the legacy C# sync implementation as option when calling `connect()`.